v:0.1.8.0.2 style: small css edits; feature: setting modal and separate window of ingredient available via direct url;
