+++ src/components/common/Modal/ModalStyles.module.scss
@@ -1,3 +1,5 @@
+@import 'styles/theme_colors';
+
 .modal_overlay {
   position: fixed;
   z-index: 100;
@@ -9,6 +11,7 @@
   left: 0;
   bottom: 0;
   right: 0;
+
 
   background-color: #000000;
   opacity: .6;
@@ -24,16 +27,24 @@
   z-index: 200;
   top: 50%;
   left: 50%;
+
+  transition: .5s all;
   -webkit-transform: translate(-50%, -50%);
   transform: translate(-50%, -50%);
 
   width: 720px;
 
   svg {
+    transition: .5s all;
     cursor: pointer;
     position: absolute;
+    z-index: 300;
     right: 40px;
-    top: 80px;
+    top: -280px;
+  }
+
+  svg:hover {
+    fill: $scroll_active;
   }
 }
