+++ src/pages/IngredientDetailsPage/IngredientDetailsStyles.module.scss
@@ -4,6 +4,8 @@
   height: 640px;
   width: 720px;
 
+
+
   position: fixed;
   left: calc(50% - 360px);
   top: calc(50% - 320px);
@@ -55,4 +57,41 @@
     }
   }
 }
+
+.fade-in-image {
+  animation: fadeIn 5s;
+  -webkit-animation: fadeIn 5s;
+  -moz-animation: fadeIn 5s;
+  -o-animation: fadeIn 5s;
+  -ms-animation: fadeIn 5s;
+}
+
+
+.fadeIn {
+  animation: fadeIn .3s;
+}
+
+.fadeOut {
+  animation: fadeOut .3s;
+}
+
+@keyframes fadeIn {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeOut {
+  100% {
+    opacity: 1;
+  }
+  0% {
+    opacity: 0;
+  }
+}
+
+
