+++ src/pages/index.ts
@@ -4,8 +4,7 @@ import NotFound404 from "pages/NotFound404.tsx";
 import RegisterPage from "pages/RegisterPage.tsx";
 import ForgotPage from "pages/ForgotPage.tsx";
 import ProfilePage from "pages/ProfilePage.tsx"
-import IngredientDetailsModal from "pages/IngredientDetailsPage/IngredientDetailsModal.tsx"
-import IngredientDetailsDefault from "pages/IngredientDetailsPage/IngredientDetailsDefault.tsx"
+import IngredientDetails from "pages/IngredientDetailsPage/IngredientDetails.tsx"
 
 export {
     LoginPage,
@@ -14,6 +13,5 @@ export {
     ForgotPage,
     RegisterPage,
     ProfilePage,
-    IngredientDetailsModal,
-    IngredientDetailsDefault
+    IngredientDetails,
 };
