+++ src/services/slices/ingredientsSlice.ts
@@ -4,33 +4,34 @@ import {checkResponse} from 'utils/check-response';
 import {IIngredientsListSlice} from "interfaces/sliceInterfaces";
 
 export const fetchIngredients = createAsyncThunk('ingredientsListSlice/fetchIngredients', async () => {
-	const response = await fetch(`${BASE_URL}/ingredients`);
-	const data = await checkResponse(response);
-	return data.data;
+    const response = await fetch(`${BASE_URL}/ingredients`);
+    const data = await checkResponse(response);
+    return data.data;
 });
 
+
 const ingredientsListSlice = createSlice({
-	name: 'ingredientsListSlice',
-	initialState: {
-		ingredients: [],
-		status: 'idle',
-		error: null,
-	} as IIngredientsListSlice,
-	reducers: {},
-	extraReducers: (builder) => {
-		builder
-			.addCase(fetchIngredients.pending, (state) => {
-				state.status = 'loading';
-			})
-			.addCase(fetchIngredients.fulfilled, (state, action) => {
-				state.status = 'succeeded';
-				state.ingredients = action.payload;
-			})
-			.addCase(fetchIngredients.rejected, (state, action) => {
-				state.status = 'failed';
-				state.error = action.error.message !== undefined ? action.error.message : null;
-			});
-	},
+    name: 'ingredientsListSlice',
+    initialState: {
+        ingredients: [],
+        status: 'idle',
+        error: null,
+    } as IIngredientsListSlice,
+    reducers: {},
+    extraReducers: (builder) => {
+        builder
+            .addCase(fetchIngredients.pending, (state) => {
+                state.status = 'loading';
+            })
+            .addCase(fetchIngredients.fulfilled, (state, action) => {
+                state.status = 'succeeded';
+                state.ingredients = action.payload;
+            })
+            .addCase(fetchIngredients.rejected, (state, action) => {
+                state.status = 'failed';
+                state.error = action.error.message !== undefined ? action.error.message : null;
+            });
+    },
 });
 
 export default ingredientsListSlice.reducer;
